Game states and enemy states arranged.  Game data manager created for managing datas.

+++ Assets/Scripts/Enemy/EnemyData.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "EnemyData.asset", menuName = "Game Datas/Enemy Data")]
 public class EnemyData : ScriptableObject
 {
-    public float Health;
-    public float Damage;
+    public int Health;
+    public int Damage;
     public int CoinAward;
     public float ProjectileSpeed;
 }
